refactor: move arrow convert function into file
