+++ rust/quary-databases/src/lib.rs
@@ -5,3 +5,4 @@ pub mod databases_postgres;
 pub mod databases_redshift;
 mod databases_snowflake;
 mod databases_sqlite;
+mod database_arrow_helper;
